pyproject.toml
[tool.poetry]
name = "fast_wheels"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.61.2"
pymongo = "^3.11.1"
motor = "^2.3.0"
uvicorn = "^0.12.2"
Scrapy = "^2.4.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
